[Static Code Analysis] [6/?] Multithreading pass

+++ Source/Engine/Systems/ResourceManager/Resource/Resource.cpp
@@ -167,7 +167,7 @@ std::string Resource::GetLibraryPath() const
 
 int16 Resource::GetIndexFromType(const ResourceType& type)
 {
-	return static_cast<int16>(std::to_underlying(type));
+	return std::to_underlying(type);
 }
 
 std::string Resource::GetLibraryExtensionFromType(const ResourceType type)
